refactor: move other types into `./types.ts`
