+++ v-next/hardhat-ignition-ethers/src/internal/ethers-ignition-helper.ts
@@ -1,5 +1,8 @@
 import "@ignored/hardhat-vnext-ethers";
-import type { EthersIgnitionHelper } from "../types.js";
+import type {
+  EthersIgnitionHelper,
+  IgnitionModuleResultsTToEthersContracts,
+} from "../types.js";
 import type { ArtifactManager } from "@ignored/hardhat-vnext/types/artifacts";
 import type { HardhatConfig } from "@ignored/hardhat-vnext/types/config";
 import type {
@@ -14,8 +17,6 @@ import type {
   Future,
   IgnitionModule,
   IgnitionModuleResult,
-  NamedArtifactContractAtFuture,
-  NamedArtifactContractDeploymentFuture,
   StrategyConfig,
   SuccessfulDeploymentResult,
 } from "@ignored/hardhat-vnext-ignition-core";
@@ -39,21 +40,6 @@ import {
   deploy,
   isContractFuture,
 } from "@ignored/hardhat-vnext-ignition-core";
-
-export type IgnitionModuleResultsTToEthersContracts<
-  ContractNameT extends string,
-  IgnitionModuleResultsT extends IgnitionModuleResult<ContractNameT>,
-> = {
-  [contract in keyof IgnitionModuleResultsT]: IgnitionModuleResultsT[contract] extends
-    | NamedArtifactContractDeploymentFuture<ContractNameT>
-    | NamedArtifactContractAtFuture<ContractNameT>
-    ? TypeChainEthersContractByName<ContractNameT>
-    : Contract;
-};
-
-// TODO: Make this work to have support for TypeChain
-// eslint-disable-next-line @typescript-eslint/no-unused-vars -- this is a placeholder for TypeChain support
-export type TypeChainEthersContractByName<ContractNameT> = Contract;
 
 export class EthersIgnitionHelperImpl<ChainTypeT extends ChainType | string>
   implements EthersIgnitionHelper
